+++ mirtop/gff/body.py
@@ -1,10 +1,13 @@
 from collections import defaultdict, OrderedDict
+from mirtop.mirna.realign import get_mature_sequence, align_from_variants, variant_to_5p, variant_to_3p, variant_to_add, read_id
+from mirtop.mirna import fasta, mapper
 
 import mirtop.libs.logger as mylog
 logger = mylog.getLogger(__name__)
 
-def create(reads, database, sample):
+def create(reads, database, sample, args):
     """Read https://github.com/miRTop/mirtop/issues/9"""
+    sep = " " if args.out_format == "gtf" else "="
     seen = set()
     lines = defaultdict(defaultdict)
     seen_ann = {}
@@ -13,6 +16,9 @@ def create(reads, database, sample):
     n_hits = 0
     n_reads = 0
     n_seen = 0
+    if args.add_extra:
+        precursors = args.precursors
+        matures = args.matures
     for r, read in reads.iteritems():
         hits = set()
         [hits.add(mature.mirna) for mature in read.precursors.values() if mature.mirna]
@@ -38,23 +44,31 @@ def create(reads, database, sample):
                 strand = iso.strand
                 start, end = iso.start, iso.end
                 score = iso.map_score
-                filter = iso.filter
                 mirName = iso.mirna
                 preName = p
                 Variant = iso.formatGFF()
                 Cigar = iso.cigar
                 counts = read.counts
                 Filter = iso.filter
-                attrb = ("Read {r}; UID {idseq}; Name {mirName}; Parent {preName}; Variant {Variant}; Cigar {Cigar}; Expression {counts}; Filter {Filter}; Hits {hits};").format(**locals())
-                res = ("{chrom}\t{database}\t{source}\t{start}\t{end}\t{score}\t{strand}\t.\t{attrb}").format(**locals())
+
+                attrb = ("Read {r}; UID {idseq}; Name {mirName}; Parent {preName};"
+                        " Variant {Variant}; Cigar {Cigar}; Expression {counts};"
+                        " Filter {Filter}; Hits {hits};").format(**locals())
+                line = ("{chrom}\t{database}\t{source}\t{start}\t{end}\t{score}\t{strand}\t.\t{attrb}").format(**locals())
+                logger.debug("GFF::%s" % line)
+                if args.add_extra:
+                    extra = variant_with_nt(line, precursors, matures)
+                    line = "%s Changes %s;" % (line, extra)
+
+                line = paste_columns(read_gff_line(line), sep = sep)
                 if annotation in seen_ann and seq.find("N") < 0 and seen_ann[annotation].split("\t")[0].find("N") < 0:
                     logger.warning("Same isomir %s from different sequence: \n%s and \n%s" % (annotation, res, seen_ann[annotation]))
-                seen_ann[annotation] = res
+                seen_ann[annotation] = line
                 logger.debug("GFF::external %s" % iso.external)
                 if start not in lines[chrom]:
                     lines[chrom][start] = []
-                lines[chrom][start].append([annotation, chrom, counts, sample, res])
-                logger.debug("GFF::%s" % res)
+                lines[chrom][start].append([annotation, chrom, counts, sample, line])
+                logger.debug("GFF::%s" % line)
                 n_hits += 1
             else:
                 n_seen += 1
@@ -110,8 +124,8 @@ def read_gff_line(line):
         return line
     cols = line.strip().split("\t")
     sep = guess_format(line)
-    if len(cols) != 9:
-        raise ValueError("Line doesn't have 9 elements: %s" % line)
+    if len(cols) < 9:
+        raise ValueError("Line has less than 9 elements: %s" % line)
     fields = {'chrom': cols[0],
               'source': cols[1],
               'type': cols[2],
@@ -122,3 +136,34 @@ def read_gff_line(line):
               'ext': cols[7],
               'attrb': read_attributes(line, sep)}
     return fields
+
+def variant_with_nt(line, precursors, matures):
+    """
+    Return nucleotides changes for each variant type
+    using Variant attribute, precursor sequences and
+    mature position.
+    """
+    cols = read_gff_line(line)
+    attr = cols["attrb"]
+    read = read_id(attr["UID"])
+    logger.debug("GFF::BODY::precursors %s" % precursors[attr["Parent"]])
+    logger.debug("GFF:BODY::mature %s" % matures[attr["Parent"]][attr["Name"]])
+    t5 = variant_to_5p(precursors[attr["Parent"]],
+                       matures[attr["Parent"]][attr["Name"]],
+                       attr["Variant"])
+    t3 = variant_to_3p(precursors[attr["Parent"]],
+                       matures[attr["Parent"]][attr["Name"]],
+                       attr["Variant"])
+    add = variant_to_add(read,
+                         attr["Variant"])
+    mature_sequence = get_mature_sequence(precursors[attr["Parent"]],
+                                          matures[attr["Parent"]][attr["Name"]])
+    logger.debug("GFF::BODY::mature_sequence %s" % mature_sequence)
+    mm = align_from_variants(read,
+                             mature_sequence,
+                             attr["Variant"])
+    if len(mm) > 1:
+        mm = "".join(["".join(map(str, m)) for m in mm])
+    else:
+        mm = "0"
+    return "iso_5p:%s,iso_3p:%s,iso_add:%s,iso_snp:%s" % (t5, t3, add, mm)